refactor(common): Move widget renderer into dashboard folder
