+++ packages/telestion-client-common/src/components/pages/dashboard-page/dashboard/dashboard.tsx
@@ -3,7 +3,7 @@ import { Grid, View } from '@adobe/react-spectrum';
 import { Dashboard as DashboardType } from '@wuespace/telestion-client-types';
 
 import { OverflowFix } from './overflow-fix';
-import { WidgetRenderer } from '../widget-renderer/widget-renderer';
+import { WidgetRenderer } from './widget-renderer/widget-renderer';
 
 /**
  * React Props of {@link OverflowFix}
