+++ packages/telestion-client-common/src/components/pages/dashboard-page/dashboard/widget-renderer/widget-renderer.tsx
@@ -6,7 +6,7 @@ import { WidgetDefinition } from '@wuespace/telestion-client-types';
 import NotFound from '@spectrum-icons/illustrations/NotFound';
 
 import { WidgetErrorMessage } from './widget-error-message';
-import { useWidgets } from '../../../contexts/widgets-context';
+import { useWidgets } from '../../../../contexts/widgets-context';
 
 function ErrorFallback({ error, resetErrorBoundary }: FallbackProps) {
 	return (
